Move the MARC 008 parsers out of internal_model and into sierra_transformer
